Move files to appropriate modules
