+++ gobblin-service/src/main/java/org/apache/gobblin/service/monitoring/GitConfigListener.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.gobblin.service.modules.core;
+package org.apache.gobblin.service.monitoring;
 
 import com.google.common.collect.Sets;
 import com.google.common.io.Files;
